+++ i-study-app/src/screens/ManageUser/index.js
@@ -4,38 +4,74 @@ import Row from "react-bootstrap/Row";
 import Form from "react-bootstrap/Form";
 import Col from "react-bootstrap/Col";
 import Table from "react-bootstrap/Table";
+import Pagination from "react-bootstrap/Pagination";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteUser, getUserList, updateUserProfile } from "../../thunks/user";
 import * as userSelectors from "../../selectors/user";
 import { USER_TYPE_NAME, USER_TYPE_NUMBER } from "../../utilities/constants";
 import Button from "react-bootstrap/esm/Button";
-import { Link } from "react-router-dom";
+import { Link, Navigate, useNavigate } from "react-router-dom";
 import EditModal from "./components/EditModal";
+import CommonFormGroup from "../common/CommonFormGroup";
+import { IoAddCircleOutline } from "react-icons/io5";
 
 const ManageUser = () => {
-  const [showModal, setShowModal] = useState(false);
-  const [editingUser, setEditingUser] = useState(null);
-  const [userType, setUserType] = useState("");
-  const [displayedUserList, setDisplayedUserList] = useState([]);
   const dispatch = useDispatch();
+  const navigate = useNavigate();
+
   const user = useSelector((state) => state.user);
   const userList = userSelectors.getUserList(user);
 
+  const [showModal, setShowModal] = useState(false);
+  const [editingUser, setEditingUser] = useState(null);
+  const [userType, setUserType] = useState("");
+  const [searchText, setSearchText] = useState("");
+  const [filteredUserList, setFilteredUserList] = useState([]);
+  const [displayedUserList, setDisplayedUserList] = useState([]);
+  const [maxPage, setMaxPage] = useState(0);
+  const [currentPage, setCurrentPage] = useState(0);
+  const [timeoutId, setTimeoutId] = useState(0);
+
   useEffect(() => {
+    setMaxPage(0);
+    setCurrentPage(0);
+    setUserType("");
     dispatch(getUserList());
   }, [dispatch]);
 
   useEffect(() => {
     setShowModal(false);
-    setDisplayedUserList(
-      userList.filter((user) =>
-        userType === "" ? true : userSelectors.getUserType(user) === userType
-      )
+    let filteredUserListCopy = userList.filter((user) =>
+      userType === "" ? true : userSelectors.getUserType(user) === userType
     );
+    setMaxPage(0);
+    setCurrentPage(0);
+    setFilteredUserList(filteredUserListCopy);
+    setMaxPage(Math.ceil(filteredUserListCopy.length / 10));
+    setDisplayedUserList(filteredUserListCopy.slice(0, 10));
   }, [userList, userType]);
 
-  const onClickDelete = (userId) => {
-    dispatch(deleteUser({ userId }));
+  useEffect(
+    () =>
+      setDisplayedUserList(
+        filteredUserList.slice(currentPage * 10, currentPage * 10 + 10)
+      ),
+    [currentPage]
+  );
+
+  useEffect(() => {
+    if (timeoutId) clearTimeout(timeoutId);
+    setTimeoutId(
+      setTimeout(() => {
+        dispatch(getUserList({ name: searchText }));
+      }, 1000)
+    );
+  }, [searchText]);
+
+  const onClickDelete = (name, userId) => {
+    if (window.confirm(`Are you sure to delete the user named ${name}?`))
+      dispatch(deleteUser({ userId }));
+    // dispatch(deleteUser({ userId }));
   };
 
   const onClickEdit = (element) => {
@@ -49,58 +85,68 @@ const ManageUser = () => {
 
   return (
     <>
-      <div className="mx-auto my-4 w-75">
-        <Row>
-          <Col>
-            <h3 className="my-2">Manage User</h3>
-          </Col>
-          <Col />
-          <Col xs sm="auto">
-            <Button variant="primary" as={Link} to="../register_user">
-              Register User
-            </Button>
-          </Col>
-        </Row>
+      <div className="pagetitle">
+        <h1>Manage User</h1>
+        <nav>
+          <ol className="breadcrumb">
+            <li className="breadcrumb-item active">User</li>
+          </ol>
+        </nav>
+      </div>
 
-        <Card className="py-5 px-5 justify-content-flex-start">
-          <Row>
-            <Col lg="auto">
-              <Form.Control type="text" placeholder="Search.." />
-            </Col>
-            <Col />
-            <Col lg="auto">
-              <Form.Group as={Row} className="d-flex align-items-center">
-                <Col>Filter:</Col>
-                <Col lg="auto">
-                  <Form.Select
-                    value={userType}
-                    onChange={(e) =>
-                      setUserType(parseInt(e.target.value) || "")
-                    }
-                  >
-                    <option value="">All User</option>
-                    <option value={USER_TYPE_NUMBER["TEACHER"]}>Teacher</option>
-                    <option value={USER_TYPE_NUMBER["STUDENT"]}>Student</option>
-                  </Form.Select>
-                </Col>
-              </Form.Group>
-            </Col>
-          </Row>
-          <Table className="my-4" bordered>
+      <Card className="card">
+        <Card.Body className="card-body pt-4">
+          <div className="d-flex justify-content-between">
+            <div>
+              <Form.Control
+                type="text"
+                placeholder="Search.."
+                style={{ width: "400px" }}
+                value={searchText}
+                onChange={(e) => setSearchText(e.target.value)}
+              />
+            </div>
+            <CommonFormGroup as={Row} className="d-flex align-items-center">
+              <Col>Filter:</Col>
+              <Col lg="auto">
+                <Form.Select
+                  value={userType}
+                  onChange={(e) => setUserType(parseInt(e.target.value) || "")}
+                >
+                  <option value="">All User</option>
+                  <option value={USER_TYPE_NUMBER["TEACHER"]}>Teacher</option>
+                  <option value={USER_TYPE_NUMBER["STUDENT"]}>Student</option>
+                </Form.Select>
+              </Col>
+            </CommonFormGroup>
+          </div>
+          <div className="d-flex flex-row justify-content-end">
+            <div
+              className="d-flex flew-row align-items-center justify-content-center"
+              style={{ color: "blue", cursor: "pointer" }}
+              onClick={() => {
+                navigate("./register_user");
+              }}
+            >
+              <IoAddCircleOutline style={{ margin: "0 8px" }} />
+              <b>Add New User</b>
+            </div>
+          </div>
+          <Table className="my-2" bordered>
             <thead>
               <tr>
                 <th className="text-center">No.</th>
                 <th className="text-center">Name</th>
                 <th className="text-center">User Type</th>
                 {/* <th>Created Date</th> */}
-                <th className="text-center">Option</th>
+                <th className="text-center">Operation</th>
               </tr>
             </thead>
             <tbody>
               {displayedUserList.map((element, index) => {
                 return (
-                  <tr key={userSelectors.getUserId(element)}>
-                    <td>{index + 1}</td>
+                  <tr key={userSelectors.getUserId(element)} className="py-2">
+                    <td>{currentPage * 10 + index + 1}</td>
                     <td>{userSelectors.getName(element)}</td>
                     <td>
                       {USER_TYPE_NAME[userSelectors.getUserType(element)]}
@@ -116,7 +162,10 @@ const ManageUser = () => {
                           <Button
                             variant="danger"
                             onClick={() =>
-                              onClickDelete(userSelectors.getUserId(element))
+                              onClickDelete(
+                                userSelectors.getName(element),
+                                userSelectors.getUserId(element)
+                              )
                             }
                           >
                             Delete
@@ -129,8 +178,19 @@ const ManageUser = () => {
               })}
             </tbody>
           </Table>
-        </Card>
-      </div>
+          <Pagination>
+            {[...Array(maxPage).keys()].map((e) => (
+              <Pagination.Item
+                key={e}
+                active={e === currentPage}
+                onClick={() => setCurrentPage(e)}
+              >
+                {e + 1}
+              </Pagination.Item>
+            ))}
+          </Pagination>
+        </Card.Body>
+      </Card>
       <EditModal
         showModal={showModal}
         onCloseModal={() => setShowModal(false)}
